feat: product filter in article landing EXLM-1542
